selenium-pytest-bdd-allure-qtest-ci/requirements.txt
pytest==8.3.2
pytest-bdd==7.1.2
selenium==4.23.1
webdriver-manager==4.0.2
allure-pytest==2.13.5
requests==2.32.3
python-dotenv==1.0.1

pytest-xdist==3.6.1
